Refactor protocol manager and models for Python alignment
- Rename model_selection.go to protocol_manager.go and update types to match Python enums and structures - Make CostBias
in ModelSelectionRequest a pointer for optionality - Update minion alternative handling to use TaskType instead of
Provider - Simplify parameter service default values and logging - Update protocol manager to handle nil CostBias with
default value
