+++ adaptive-backend/internal/services/chat/completions/parameter_service.go
@@ -25,21 +25,22 @@ func (s *ParameterService) ApplyModelParameters(
 
 	// Apply parameters with logging
 	req.MaxTokens = params.MaxTokens
-	fiberlog.Debugf("[%s] Applied MaxTokens: %d", requestID, params.MaxTokens)
+	fiberlog.Debugf("[%s] Applied MaxTokens: %v", requestID, params.MaxTokens)
 
 	req.Temperature = params.Temperature
-	fiberlog.Debugf("[%s] Applied Temperature: %f", requestID, params.Temperature)
+	fiberlog.Debugf("[%s] Applied Temperature: %v", requestID, params.Temperature)
 
 	req.TopP = params.TopP
-	fiberlog.Debugf("[%s] Applied TopP: %f", requestID, params.TopP)
+	fiberlog.Debugf("[%s] Applied TopP: %v", requestID, params.TopP)
+
 	req.PresencePenalty = params.PresencePenalty
-	fiberlog.Debugf("[%s] Applied PresencePenalty: %f", requestID, params.PresencePenalty)
+	fiberlog.Debugf("[%s] Applied PresencePenalty: %v", requestID, params.PresencePenalty)
 
 	req.FrequencyPenalty = params.FrequencyPenalty
-	fiberlog.Debugf("[%s] Applied FrequencyPenalty: %f", requestID, params.FrequencyPenalty)
+	fiberlog.Debugf("[%s] Applied FrequencyPenalty: %v", requestID, params.FrequencyPenalty)
 
 	req.N = params.N
-	fiberlog.Debugf("[%s] Applied N: %d", requestID, params.N)
+	fiberlog.Debugf("[%s] Applied N: %v", requestID, params.N)
 
 	fiberlog.Infof("[%s] Parameter application complete", requestID)
 	return nil
@@ -47,20 +48,13 @@ func (s *ParameterService) ApplyModelParameters(
 
 // GetDefaultParameters returns sensible default parameters for chat completions
 func (s *ParameterService) GetDefaultParameters() models.OpenAIParameters {
-	temperature := 0.7
-	topP := 1.0
-	maxTokens := 1000
-	n := 1
-	presencePenalty := 0.0
-	frequencyPenalty := 0.0
-
 	return models.OpenAIParameters{
-		Temperature:      param.Opt[float64]{Value: temperature},
-		TopP:             param.Opt[float64]{Value: topP},
-		MaxTokens:        param.Opt[int64]{Value: int64(maxTokens)},
-		N:                param.Opt[int64]{Value: int64(n)},
-		PresencePenalty:  param.Opt[float64]{Value: presencePenalty},
-		FrequencyPenalty: param.Opt[float64]{Value: frequencyPenalty},
+		Temperature:      param.Opt[float64]{Value: 0.7},
+		TopP:             param.Opt[float64]{Value: 1.0},
+		MaxTokens:        param.Opt[int64]{Value: 1000},
+		N:                param.Opt[int64]{Value: 1},
+		PresencePenalty:  param.Opt[float64]{Value: 0.0},
+		FrequencyPenalty: param.Opt[float64]{Value: 0.0},
 	}
 }
